Fix bug
The url used in axios request is a relative url to the base url, so we don't need the "'/' +".
This caused an issue of receiving always the Http code 302. After diagnost the issue i found that the Route::post to "queries" for example never reached (I did a test using dd()). But when i removed the '/' from this file all works fine. (At this point you should disable the CSRF middleware to get thing working waiting for fixing the 419 HTTP code received if the Csrf middleware is enabled in Kernel.php).