Allow for multi-tenancy
Changed the "show" method to allow for multi-tenancy applications to still utilise the package. Applications which use Route Model Binding for subdomains aren't taken into account by this method and instead the $id would return the tenant model. By getting the entry id from the request, it retains the functionality and allows for multi-tenant applications to use this package per tenant.